requirements.txt
azure-functions==1.24.0
azure-identity==1.23.0
azure-mgmt-network==29.0.0
azure-storage-blob==12.27.0
python-dateutil==2.8.2
requests==2.31.0
